Added This Error, Fixed Errors into a single Error Type. Split the Code into Seperate Files
